Moved import_shared_memory to wasi specific, and use this function on the run/wasi cli command directly
Fixed linter

Fixed js build

Fixed linter
