+++ lib/wasi/src/utils.rs
@@ -1,5 +1,7 @@
 use std::collections::BTreeSet;
-use wasmer::Module;
+#[cfg(not(feature = "js"))]
+use wasmer::vm::VMSharedMemory;
+use wasmer::{AsStoreMut, Imports, Memory, Module};
 use wasmer_wasi_types::wasi::Errno;
 
 #[allow(dead_code)]
@@ -46,6 +48,44 @@ pub fn map_io_err(err: std::io::Error) -> Errno {
         ErrorKind::Unsupported => Errno::Notsup,
         _ => Errno::Io,
     }
+}
+
+/// Imports (any) shared memory into the imports.
+/// (if the module does not import memory then this function is ignored)
+#[cfg(not(feature = "js"))]
+pub fn wasi_import_shared_memory(
+    imports: &mut Imports,
+    module: &Module,
+    store: &mut impl AsStoreMut,
+) {
+    // Determine if shared memory needs to be created and imported
+    let shared_memory = module
+        .imports()
+        .memories()
+        .next()
+        .map(|a| *a.ty())
+        .map(|ty| {
+            let style = store.as_store_ref().tunables().memory_style(&ty);
+            VMSharedMemory::new(&ty, &style).unwrap()
+        });
+
+    if let Some(memory) = shared_memory {
+        // if the memory has already be defined, don't redefine it!
+        if !imports.exists("env", "memory") {
+            imports.define(
+                "env",
+                "memory",
+                Memory::new_from_existing(store, memory.into()),
+            );
+        }
+    };
+}
+#[cfg(feature = "js")]
+pub fn wasi_import_shared_memory(
+    _imports: &mut Imports,
+    _module: &Module,
+    _store: &mut impl AsStoreMut,
+) {
 }
 
 /// The version of WASI. This is determined by the imports namespace
